// types/author.ts
export interface Author {
  id: number;
  user_id: null | number;
  name: string;
  slug: string;
  pseudonym: null | string;
  alias: null | string;
  bio: null | string;
  avatar: null | string;
  email: null | string;
  nationality: null | string;
  birthday: null | string;
  social_media: null | string;
  created_at: string;
  updated_at: string;
}